+++ pkg/client/service.go
@@ -204,242 +204,6 @@ func (c *Client) SearchForItems(itemName string) ([]*svc.File, []*svc.Directory,
 	return files, dirs, nil
 }
 
-// ------ configuration --------------------------------
-
-// update user-specific settings
-func (c *Client) UpdateConfigSetting(setting, value string) error {
-	switch setting {
-	case "CLIENT_NAME":
-		return c.updateClientName(value)
-	case "CLIENT_USERNAME": // TODO
-		return c.updateUserAlias(value)
-	case "CLIENT_EMAIL":
-		return c.updateClientEmail(value)
-	case "CLIENT_PASSWORD":
-		return c.updateUserPassword(c.User.Password, value)
-	case "CLIENT_PORT":
-		return c.updateClientPort(value)
-	case "CLIENT_BACKUP_DIR":
-		return c.UpdateBackupPath(value)
-	case "CLIENT_LOCAL_BACKUP":
-		return c.SetLocalBackup(value)
-	case "CLIENT_PROFILE_PIC":
-		return c.updateClientIcon(value)
-	case "CLIENT_NEW_SERVICE":
-		return envCfgs.Set(setting, value)
-	case "NEW_SERVICE":
-		return envCfgs.Set(setting, value)
-	default:
-		return fmt.Errorf("unsupported setting: '%s'", setting)
-	}
-}
-
-// enable or disable backing up files to local storage.
-func (c *Client) SetLocalBackup(modeStr string) error {
-	mode, err := strconv.ParseBool(modeStr)
-	if err != nil {
-		c.log.Error(fmt.Sprintf("failed to parse string: %v", err))
-		return err
-	}
-	c.Conf.LocalBackup = mode
-	if err := envCfgs.Set("CLIENT_LOCAL_BACKUP", strconv.FormatBool(mode)); err != nil {
-		return err
-	}
-	if err := c.SaveState(); err != nil {
-		c.log.Error("failed to update state file: " + err.Error())
-	} else {
-		if mode {
-			c.log.Info("local backup enabled")
-		} else {
-			c.log.Info("local backup disabled")
-		}
-	}
-	return nil
-}
-
-// update the backup configurations for the service and all client-side
-// files and directories
-func (c *Client) UpdateBackupPath(newDirPath string) error {
-	if !c.isDirPath(newDirPath) {
-		c.log.Error("path is not a directory")
-		return fmt.Errorf("path is not a directory")
-	}
-	if c.LocalBackupDir == newDirPath {
-		return nil // nothing to replace
-	}
-	return c.updateBackupPaths(filepath.Clean(newDirPath))
-}
-
-// If a user specifies a custom backup directory, then all
-// local backup items need to have their backup paths updated.
-// we don't want files to be backed up to a default directory if
-// a user tries to specify otherwise.
-func (c *Client) updateBackupPaths(newPath string) error {
-	oldPath := c.Root
-	dirs := c.Drive.GetDirs()
-	for _, dir := range dirs {
-		dir.BackupPath = strings.Replace(dir.BackupPath, oldPath, newPath, 1)
-	}
-	if err := c.Db.UpdateDirs(dirs); err != nil {
-		return err
-	}
-	files := c.Drive.GetFiles()
-	for _, file := range files {
-		file.BackupPath = strings.Replace(file.BackupPath, oldPath, newPath, 1)
-	}
-	if err := c.Db.UpdateFiles(files); err != nil {
-		return err
-	}
-	c.Conf.BackupDir = newPath
-	if err := envCfgs.Set("CLIENT_BACKUP_DIR", newPath); err != nil {
-		return err
-	}
-	if err := c.SaveState(); err != nil {
-		return err
-	}
-	return nil
-}
-
-// update user's name
-func (c *Client) updateClientName(newName string) error {
-	if newName == c.Conf.User && newName == c.User.Name && newName == c.Drive.OwnerName {
-		return nil
-	}
-	user, err := c.Db.GetUser(c.UserID)
-	if err != nil {
-		return err
-	}
-	if user == nil {
-		return fmt.Errorf("user not found: %v", c.UserID)
-	}
-	c.User.Name = newName
-	c.Conf.User = newName
-	c.Drive.OwnerName = newName
-	user.Name = newName
-	if err := c.Db.UpdateUser(user); err != nil {
-		return err
-	}
-	if err := c.Db.UpdateDrive(c.Drive); err != nil {
-		return err
-	}
-	if err := envCfgs.Set("CLIENT_NAME", newName); err != nil {
-		return err
-	}
-	// TODO: sync with remote server, if necessary
-	if err := c.SaveState(); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *Client) updateUserAlias(newAlias string) error {
-	if newAlias == c.User.UserName || newAlias == "" {
-		return nil
-	}
-	c.User.UserName = newAlias
-	c.Conf.UserAlias = newAlias
-	if err := c.Db.UpdateUser(c.User); err != nil {
-		return err
-	}
-	if err := envCfgs.Set("CLIENT_USERNAME", newAlias); err != nil {
-		return err
-	}
-	// TODO: sync with remote server, if necessary
-	if err := c.SaveState(); err != nil {
-		return err
-	}
-	return nil
-}
-
-// TODO: update all items owner name in the DB with the new user's new name
-func (c *Client) updateFileOwnerName(oldName, newName string) error {
-	if newName == "" {
-		return fmt.Errorf("no name provided")
-	}
-	if newName == oldName {
-		return nil
-	}
-	return nil
-}
-
-// update the value for CLIENT_PROFILE_PIC. ususally a file name.
-func (c *Client) updateClientIcon(fileName string) error {
-	if fileName == "" {
-		return fmt.Errorf("no path specified")
-	}
-	c.Conf.ProfilePic = fileName
-	return envCfgs.Set("CLIENT_PROFILE_PIC", fileName)
-}
-
-// update user's email
-func (c *Client) updateClientEmail(newEmail string) error {
-	user, err := c.Db.GetUser(c.UserID)
-	if err != nil {
-		return err
-	}
-	if user == nil {
-		return fmt.Errorf("user not found: '%s' (id=%v)", c.Conf.User, c.UserID)
-	}
-	c.User.Email = newEmail
-	user.Email = newEmail
-	if err := c.Db.UpdateUser(user); err != nil {
-		return err
-	}
-	if err := envCfgs.Set("CLIENT_EMAIL", newEmail); err != nil {
-		return err
-	}
-	// TODO: sync with remote server, if necessary
-	if err := c.SaveState(); err != nil {
-		return err
-	}
-	return nil
-}
-
-// update user's password
-func (c *Client) updateUserPassword(oldPw, newPw string) error {
-	user, err := c.Db.GetUser(c.UserID)
-	if err != nil {
-		return err
-	}
-	if user == nil {
-		return fmt.Errorf("user not found: '%s' (id=%v)", c.Conf.User, c.UserID)
-	}
-	if newPw == user.Password && newPw == c.User.Password {
-		return nil // nothing to update
-	}
-	// make sure current password is valid
-	if oldPw != user.Password && oldPw != c.User.Password {
-		return fmt.Errorf("incorrect password. password not updated")
-	}
-	user.Password = newPw
-	c.User.Password = newPw
-	// TODO: hashing of user passwords should occur before saving to DB.
-	// dont save them as plaintext! or not save the PW at all?
-	if err := c.Db.UpdateUser(user); err != nil {
-		return err
-	}
-	if err := envCfgs.Set("CLIENT_PASSWORD", newPw); err != nil {
-		return err
-	}
-	if err := c.SaveState(); err != nil {
-		return err
-	}
-	return nil
-}
-
-// update client port setting
-func (c *Client) updateClientPort(pvalue string) error {
-	port, err := strconv.Atoi(pvalue)
-	if err != nil {
-		return err
-	}
-	if c.Conf.ClientPort == port {
-		return nil // nothing to do here
-	}
-	c.Conf.ClientPort = port
-	return envCfgs.Set("CLIENT_PORT", pvalue)
-}
-
 // ------ misc --------------------------------
 
 func (c *Client) GetServerRuntime() (float64, error) {
@@ -584,11 +348,11 @@ func (c *Client) GetFileByName(name string) (*svc.File, error) {
 }
 
 /*
-add a file to the service using its file path.
+add a file to the service using its absolute path.
 
 SFS can monitor files outside of the designated root directory, so
 if we add a file this way then we should automatically make a backup of it
-in the SFS server, or locally to a designated directory, depending
+in the SFS server or locally to a designated directory, depending
 on the service configurations.
 */
 func (c *Client) AddFile(filePath string) error {
@@ -661,7 +425,7 @@ func (c *Client) AddFile(filePath string) error {
 				return err
 			}
 		} else {
-			c.log.Warn(fmt.Sprintf("failed to send metadata to server: %v", resp.Status))
+			c.log.Warn(fmt.Sprintf("failed to send metadata to server. server response code: %v", resp.Status))
 		}
 	} else {
 		// make a local backup copy of the file
