+++ pkg/client/web_client.go
@@ -125,6 +125,9 @@ func newWcRouter(client *Client) *chi.Mux {
 		r.Route("/i/{fileID}", func(r chi.Router) {
 			r.Use(server.FileCtx)
 			r.Get("/", client.FilePage) // get info about a file
+			r.Route("/open-loc", func(r chi.Router) {
+				r.Get("/", client.OpenFileLocHandler)
+			})
 		})
 		r.Route("/delete", func(r chi.Router) {
 			r.Delete("/", client.RemoveFileHandler)
