+++ pkg/configs/configs.yaml
@@ -1,0 +1,34 @@
+ADMIN_MODE: "false"
+BUFFERED_EVENTS: "true"
+CLIENT_ADDRESS: "localhost:9090"
+CLIENT_BACKUP_DI: "C:\\Users\\Jay\\Desktop\\backups"
+CLIENT_CFG_FILE: "C:\\Users\\Jay\\coding-projects\\sfs\\pkg\\configs\\configs\\configs.yaml"
+CLIENT_EMAIL: "wnATwienermann.com"
+CLIENT_HOST: "localhost"
+CLIENT_ID: "57a18de6-64c8-11ef-a2d3-d85ed3e090f7"
+CLIENT_LOCAL_BACKUP: "true"
+CLIENT_LOG_DIR: "C:\\Users\\Jay\\coding-projects\\sfs\\pkg\\logger\\logs"
+CLIENT_NAME: "wienermann nugget"
+CLIENT_NEW_SERVICE: "false"
+CLIENT_PASSWORD: "skibityboo"
+CLIENT_PORT: 9090
+CLIENT_PROFILE_PIC: "default_profile_pic.jpg"
+CLIENT_ROOT: "C:\\Users\\Jay\\coding-projects\\sfs\\pkg\\client\\run"
+CLIENT_TESTING: "C:\\Users\\Jay\\coding-projects\\sfs\\pkg\\client\\testing"
+CLIENT_USERNAME: "wiener_nugs_69"
+EVENT_BUFFER_SIZE: 2
+JWT_SECRET: "secret-key-goes-here"
+NEW_SERVICE: "false"
+SERVER_ADDR: "localhost:8080"
+SERVER_ADMIN: "admin"
+SERVER_ADMIN_KEY: "derp"
+SERVER_HOST: "localhost"
+SERVER_LOG_DIR: "C:\\Users\\Jay\\coding-projects\\sfs\\pkg\\logger\\logs"
+SERVER_PORT: 8080
+SERVER_TIMEOUT_IDLE: "900s"
+SERVER_TIMEOUT_READ: "5s"
+SERVER_TIMEOUT_WRITE: "10s"
+SERVICE_ENV: "C:\\Users\\Jay\\coding-projects\\sfs\\pkg\\env\\.env"
+SERVICE_LOG_DIR: "C:\\Users\\Jay\\coding-projects\\sfs\\pkg\\logger\\logs"
+SERVICE_ROOT: "C:\\Users\\Jay\\coding-projects\\sfs\\pkg\\server\\run\\sfs_test"
+SERVICE_TEST_ROOT: "C:\\Users\\Jay\\coding-projects\\sfs\\pkg\\server\\testing"
