+++ static/scripts/settings.js
@@ -1,10 +1,22 @@
 function submitSettings() {
-  document.getElementById("loading-spinner").style.display = "block"; // Show loading spinner
+  document.getElementById("loading-spinner").style.display = "block";
+  
+  // const theme = document.getElementById("theme").value
+  const serverSync = document.getElementById("server-sync").value;
+  const backupDir = document.getElementById("local-backup-dir").value;
+  const clientPort = document.getElementById("client-port").value;
+  const syncDelay = document.getElementById("sync-delay").value;
+
+  // TODO: handle theme on the client side
+
   const settings = {
-    theme: document.getElementById("theme").value,
-    notifications: document.getElementById("notifications").checked,
-    serverSync: document.getElementById("server-sync").checked,
+    CLIENT_LOCAL_BACKUP: serverSync,
+    CLIENT_BACKUP_DIR: backupDir,
+    CLIENT_PORT: clientPort,
+    EVENT_BUFFER_SIZE: syncDelay
   };
+
+  console.log("sending settings to server: ", JSON.stringify(settings));
 
   fetch("/settings", {
     method: "POST",
@@ -13,19 +25,19 @@ function submitSettings() {
     },
     body: JSON.stringify(settings),
   })
-    .then((response) => {
-      if (!response.ok) {
-        throw new Error("Error updating settings");
-      }
-    })
-    .then((data) => {
-      console.log("Settings updated successfully");
-      document.getElementById("loading-spinner").style.display = "none";
-      alert("Settings updated successfully")
-    })
-    .catch((error) => {
-      console.error("Error:", error);
-      document.getElementById("loading-spinner").style.display = "none";
-      alert(error)
-    });
+  .then((response) => {
+    if (!response.ok) {
+      throw new Error("Error updating settings: " + JSON.stringify(response));
+    }
+  })
+  .then((data) => {
+    console.log("Settings updated successfully");
+    document.getElementById("loading-spinner").style.display = "none";
+    alert("Settings updated successfully")
+  })
+  .catch((error) => {
+    console.error("Error:", error);
+    document.getElementById("loading-spinner").style.display = "none";
+    alert(error)
+  });
 }
