+++ static/templates/settings.html
@@ -6,7 +6,7 @@
     <div id="content">
       {{template "toolbar"}}
       <div class="settings-container">
-        <h2>Application Settings</h2>
+        <h2 style="text-align: center">Application Settings</h2>
         <form id="settings-form" style="width: 500px">
           <div class="setting-item">
             <label for="theme">Theme:</label>
@@ -42,6 +42,9 @@
           </div>
           <div class="setting-item">
             <label for="client-port">Client Port</label>
+            <p id="client-port-notification">
+              NOTE: This will take effect next time the application is launched.
+            </p>
             <input
               type="number"
               id="client-port"
@@ -49,21 +52,24 @@
               placeholder="current: {{.ClientPort}}"
             />
           </div>
+          <br />
           <div class="setting-item">
             <label for="sync-delay">Sync Delay</label>
             <p id="sync-delay-instructions">
               This affects how ofter the client will attempt to synchronize with
-              the server. The current default is one sync event per
-              {{.EventBufferSize}} detected events for a given item. Increasing
-              this number will increase the time between syncronization events.
+              the server. The current setting is one sync event per
+              {{.EventBufferSize}} detected events for a given item. Default is
+              2 detections per sync event. Increasing this number will increase
+              the time between syncronization events.
             </p>
             <input
               type="number"
               id="sync-delay"
               name="sync-delay"
-              placeholder="current (default): {{.EventBufferSize}}"
+              placeholder="current: {{.EventBufferSize}}"
             />
           </div>
+          <br />
           <button type="button" id="save-settings" onclick="submitSettings()">
             Save Settings
           </button>
@@ -73,6 +79,30 @@
         </div>
       </div>
       <script src="/static/scripts/settings.js"></script>
+      <script>
+        function checkServerStatus() {
+          let serverURL = "http://" + "{{.ServerHost}}";
+          console.log("serverURL: " + serverURL);
+          const statusText = document.getElementById("status-text");
+          fetch(serverURL)
+            .then((response) => {
+              if (response.ok) {
+                statusText.textContent = "online";
+                statusText.classList.remove("offline");
+                statusText.classList.add("online");
+              } else {
+                throw new Error("Server offline");
+              }
+            })
+            .catch((error) => {
+              statusText.textContent = "offline";
+              statusText.classList.remove("online");
+              statusText.classList.add("offline");
+            });
+        }
+        document.addEventListener("DOMContentLoaded", checkServerStatus);
+        setInterval(checkServerStatus, 60000); // Check every 60 seconds
+      </script>
     </div>
   </body>
 </html>
