+++ static/templates/user.html
@@ -6,7 +6,7 @@
     <div id="content">
       {{template "toolbar"}}
       <div class="user-info-container">
-        <h2>{{.Name}}</h2>
+        <h2 style="text-align: center">{{.Name}}</h2>
         <div class="user-info-photo">
           <img
             class="user-profile-pic"
